# apps/backend/api-gateway/doc/openapi.json
{"openapi":"3.1.0","info":{"title":"Digital Twin Backend","description":"Unified backend service for sensor data management","contact":{"name":"Digital Twin Team","email":"admin@example.com"},"version":"1.0.0"},"paths":{"/api/v1/health":{"get":{"tags":["sensor-data"],"summary":"Health check","description":"Returns the health status of the backend service.","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthCheckResponse"}}}}}}},"/api/v1/readings":{"post":{"tags":["sensor-data"],"summary":"Create sensor reading","description":"Insert a new sensor reading into the database.","operationId":"create_sensor_reading_api_v1_readings_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SensorReadingCreate"}}}},"responses":{"201":{"description":"Sensor reading stored successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["sensor-data"],"summary":"List sensor readings","description":"Retrieve the most recent sensor readings ordered by timestamp descending.","operationId":"get_sensor_data_api_v1_readings_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"description":"Maximum number of sensor readings to return","default":200,"title":"Limit"},"description":"Maximum number of sensor readings to return"}],"responses":{"200":{"description":"List of sensor readings","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SensorReadingResponse"},"title":"Response Get Sensor Data Api V1 Readings Get"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["sensor-data"],"summary":"Delete all sensor readings","description":"Remove all sensor readings from the database.","operationId":"clear_sensor_data_api_v1_readings_delete","responses":{"200":{"description":"All sensor readings deleted","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/v1/live":{"get":{"tags":["sensor-data"],"summary":"Liveness probe","operationId":"liveness_check_api_v1_live_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthCheckResponse"}}}}}}},"/api/v1/ready":{"get":{"tags":["sensor-data"],"summary":"Readiness probe","operationId":"readiness_check_api_v1_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthCheckResponse"}}}}}}}},"components":{"schemas":{"ErrorResponse":{"properties":{"detail":{"type":"string","title":"Detail","examples":["Database connection failed"]}},"type":"object","required":["detail"],"title":"ErrorResponse","description":"Standard error response"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthCheckResponse":{"properties":{"status":{"type":"string","title":"Status","examples":["healthy"]},"service":{"type":"string","title":"Service","examples":["backend"]},"version":{"type":"string","title":"Version","examples":["1.0.0"]}},"type":"object","required":["status","service","version"],"title":"HealthCheckResponse","description":"Health check response"},"MessageResponse":{"properties":{"message":{"type":"string","title":"Message","examples":["Sensor reading stored successfully"]}},"type":"object","required":["message"],"title":"MessageResponse","description":"Generic success response"},"SensorReadingCreate":{"properties":{"timestamp":{"type":"string","format":"date-time","title":"Timestamp","description":"Timestamp when the reading was captured","examples":["2026-04-23T18:30:00Z"]},"temperature":{"type":"number","maximum":150.0,"minimum":-50.0,"title":"Temperature","description":"Temperature in Celsius","examples":[24.5]},"humidity":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Humidity","description":"Relative humidity percentage","examples":[61.2]}},"type":"object","required":["timestamp","temperature","humidity"],"title":"SensorReadingCreate","description":"Request model for submitting a sensor reading"},"SensorReadingResponse":{"properties":{"timestamp":{"type":"string","format":"date-time","title":"Timestamp","examples":["2026-04-23T18:30:00Z"]},"temperature":{"type":"number","title":"Temperature","examples":[24.5]},"humidity":{"type":"number","title":"Humidity","examples":[61.2]}},"type":"object","required":["timestamp","temperature","humidity"],"title":"SensorReadingResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}